Revert "WRN-18652: Make drawing controls optional (#1208)"
This reverts commit 077cd6ad58a58cee4543f73fd2f98f606b711e84.
