+++ Drawing/Drawing.js
@@ -1,4 +1,4 @@
-/* eslint-disable react-hooks/rules-of-hooks */
+/* eslint-disable react-hooks/rules-of-hooks, react/jsx-no-bind */
 
 /**
  * Sandstone styled drawing components and behaviors.
@@ -13,19 +13,20 @@
  */
 
 import kind from '@enact/core/kind';
-import EnactPropTypes from '@enact/core/internal/prop-types';
-import ComponentOverride from '@enact/ui/ComponentOverride';
 import {Drawing as UiDrawing} from '@enact/ui/Drawing';
+import Group from '@enact/ui/Group';
 import {Cell, Column, Layout, Row} from '@enact/ui/Layout';
 import PropTypes from 'prop-types';
 import compose from 'ramda/src/compose';
-import {useEffect, useRef, useState} from 'react';
+import {useRef, useState} from 'react';
 
+import BodyText from '../BodyText';
+import Button from '../Button';
 import Scroller from '../Scroller';
 import Skinnable from '../Skinnable';
+import Slider from '../Slider';
 
-import DrawingControls from './DrawingControls';
-import DrawingUtils from './DrawingUtils';
+import ColorPicker from './ColorPicker';
 
 import css from './Drawing.module.less';
 
@@ -47,33 +48,6 @@ const DrawingBase = kind({
 	functional: true,
 
 	propTypes: /** @lends sandstone/Drawing.DrawingBase.prototype */ {
-		/**
-		 * Sets the color of brush.
-		 *
-		 * @type {String}
-		 * @default #545BCC
-		 * @public
-		 */
-		brushColor: PropTypes.string,
-
-		/**
-		 * Sets the size of brush.
-		 *
-		 * @type {Number}
-		 * @default 5
-		 * @public
-		 */
-		brushSize: PropTypes.number,
-
-		/**
-		 * Sets the color of canvas.
-		 *
-		 * @type {String}
-		 * @default #FFFFFF
-		 * @public
-		 */
-		canvasColor: PropTypes.string,
-
 		/**
 		 * Sets the height of canvas.
 		 *
@@ -101,100 +75,13 @@ const DrawingBase = kind({
 		 * @default false
 		 * @public
 		 */
-		disabled: PropTypes.bool,
-
-		/**
-		 * Overrides the default Drawing control component to support customized behaviors.
-		 *
-		 * The provided component will receive the following props from `Drawing`:
-		 *
-		 * * `brushColor` - Color used for brush
-		 * * `brushSize` - Value used for brush size
-		 * * `canvasColor` - Color of canvas
-		 * * `disabled` - Disables drawing controls
-		 * * `fillColor` - Color used when `drawingTool=fill`
-		 * * `setBrushColor` - Called when `brushColor` is changed
-		 * * `setBrushSize` - Called when `brushSize` is changed
-		 * * `setDrawingTool` - Called when `drawingTool` is changed
-		 * * `setFillColor` - Called when `fillColor` is changed
-		 *
-		 * @type {Component|Element}
-		 * @default sandstone/Drawing.DrawingControls
-		 * @public
-		 */
-		drawingControlsComponent: EnactPropTypes.componentOverride,
-
-		/**
-		 * Sets the tool of drawing.
-		 *
-		 * @type {String}
-		 * @default brush
-		 * @public
-		 */
-		drawingTool: PropTypes.string,
-
-		/**
-		 * Overrides the default Drawing utils component to support customized behaviors.
-		 *
-		 * The provided component will receive the following props from `Drawing`:
-		 *
-		 * * `backgroundImage` - Sets an image as canvas background
-		 * * `disabled` - Disables drawing utils
-		 * * `drawingRef` - Reference of drawing
-		 * * `fileInputHandler` - Called when user selects an image as canvas background
-		 * * `setBackgroundImage` - Called when the background of canvas is changed
-		 *
-		 * @type {Component|Element}
-		 * @default sandstone/Drawing.DrawingUtils
-		 * @public
-		 */
-		drawingUtilsComponent: EnactPropTypes.componentOverride,
-
-		/**
-		 * Sets the color of fill.
-		 *
-		 * @type {String}
-		 * @default #D0BB22
-		 * @public
-		 */
-		fillColor: PropTypes.string,
-
-		/**
-		 * Displays the drawing controls.
-		 *
-		 * When `true`, the drawing controls is displayed.
-		 *
-		 * @type {Boolean}
-		 * @default true
-		 * @public
-		 */
-		showDrawingControls: PropTypes.bool,
-
-		/**
-		 * Displays the drawing utils.
-		 *
-		 * When `true`, the drawing utils is displayed.
-		 *
-		 * @type {Boolean}
-		 * @default true
-		 * @public
-		 */
-		showDrawingUtils: PropTypes.bool
+		disabled: PropTypes.bool
 	},
 
 	defaultProps: {
-		brushColor: '#545BCC',
-		brushSize: 5,
-		canvasColor: '#FFFFFF',
 		canvasHeight: 800,
 		canvasWidth: 1200,
-		disabled: false,
-		drawingControlsComponent: DrawingControls,
-		drawingTool: 'brush',
-		drawingUtilsComponent: DrawingUtils,
-		fillColor: '#D0BB22',
-		showDrawingControls: true,
-		showDrawingUtils: true
+		disabled: false
 	},
 
 	handlers: {
@@ -214,6 +101,12 @@ const DrawingBase = kind({
 			} catch (err) {
 				// failing silently
 			}
+		},
+
+		handleSelect: (ev) => {
+			const e = ev?.event;
+			const setDrawingTool = ev?.setDrawingTool;
+			setDrawingTool(e.data);
 		}
 	},
 
@@ -227,103 +120,125 @@ const DrawingBase = kind({
 		publicClassNames: true
 	},
 
-	render: ({
-		brushColor,
-		brushSize,
-		canvasColor,
-		canvasHeight,
-		canvasWidth,
-		disabled,
-		drawingControlsComponent,
-		drawingTool,
-		drawingUtilsComponent,
-		fillColor,
-		fileInputHandler,
-		showDrawingControls,
-		showDrawingUtils,
-		...rest
-	}) => {
+	render: ({canvasHeight, canvasWidth, disabled, fileInputHandler, handleSelect, ...rest}) => {
 		const [backgroundImage, setBackgroundImage] = useState(null);
-		const [brushColorValue, setBrushColorValue] = useState(brushColor);
-		const [brushSizeValue, setBrushSizeValue] = useState(brushSize);
-		const [canvasColorValue, setCanvasColorValue] = useState(canvasColor);
-		const [drawingToolValue, setDrawingToolValue] = useState(drawingTool);
-		const [fillColorValue, setFillColorValue] = useState(fillColor);
+		const [brushColor, setBrushColor] = useState('#545BCC');
+		const [brushSize, setBrushSize] = useState(5);
+		const [canvasColor, setCanvasColor] = useState('#FFFFFF');
+		const [drawingTool, setDrawingTool] = useState('brush');
+		const [fillColor, setFillColor] = useState('#D0BB22');
 		const drawingRef = useRef();
 
-		useEffect(() => {
-			setBrushSizeValue(brushSize);
-		}, [brushSize]);
-
-		useEffect(() => {
-			setDrawingToolValue(drawingTool);
-		}, [drawingTool]);
-
-		useEffect(() => {
-			setBrushColorValue(brushColor);
-		}, [brushColor]);
-
-		useEffect(() => {
-			setFillColorValue(fillColor);
-		}, [fillColor]);
-
-		useEffect(() => {
-			setCanvasColorValue(canvasColor);
-		}, [canvasColor]);
+		const brushColors = ['#000000', '#FFFFFF', '#FF0000', '#00FF00', '#0000FF', '#FFFF00', '#00FFFF', '#FF00FF'];
+		const canvasColors = ['#000000', '#FFFFFF', '#FF0000', '#00FF00', '#0000FF', '#FFFF00', '#00FFFF', '#FF00FF'];
+		const drawingTools = [
+			{icon: 'edit', key: 1, tooltipText: 'brush'},
+			{icon: 'heart', key: 2, tooltipText: 'fill'},
+			{icon: 'play', key: 3, tooltipText: 'triangle'},
+			{icon: 'popupscale', key: 4, tooltipText: 'rectangle'},
+			{icon: 'newfeature', key: 5, tooltipText: 'circle'},
+			{icon: 'square', key: 6, tooltipText: 'erase'}
+		];
+		const fillColors = ['#000000', '#FFFFFF', '#FF0000', '#00FF00', '#0000FF', '#FFFF00', '#00FFFF', '#FF00FF'];
 
 		return (
 			<Scroller>
 				<Layout {...rest}>
 					<Cell className={css.toolbar} shrink size="15%">
-						{showDrawingControls ? (
-							<ComponentOverride
-								brushColor={brushColorValue}
-								brushSize={brushSizeValue}
-								canvasColor={canvasColorValue}
-								component={drawingControlsComponent}
+						<Cell>
+							<BodyText css={css} disabled={disabled}>Drawing tools</BodyText>
+							<Group
+								childComponent={Button}
+								childProp="tooltipText"
+								className={css.drawingTools}
+								defaultSelected={0}
+								itemProps={{
+									disabled: disabled,
+									size: 'small'
+								}}
+								onSelect={(event) => handleSelect({event, setDrawingTool})}
+								select={'radio'}
+								selectedProp="selected"
+							>
+								{drawingTools}
+							</Group>
+						</Cell>
+						<Cell>
+							<BodyText css={css} disabled={disabled}>Brush size</BodyText>
+							<Slider
+								backgroundProgress={0}
+								css={css}
+								defaultValue={brushSize}
 								disabled={disabled}
-								fillColor={fillColorValue}
-								setBrushColor={setBrushColorValue}
-								setBrushSize={setBrushSizeValue}
-								setCanvasColor={setCanvasColorValue}
-								setDrawingTool={setDrawingToolValue}
-								setFillColor={setFillColorValue}
+								max={30}
+								min={1}
+								onChange={(e) => {
+									setBrushSize(e.value);
+								}}
+								step={1}
+								tooltip={false}
 							/>
-						) :
-							null
-						}
+						</Cell>
+						<Cell className={css.colors}>
+							<BodyText css={css} disabled={disabled}>Colors</BodyText>
+							<ColorPicker
+								color={brushColor}
+								colorHandler={setBrushColor}
+								disabled={disabled}
+								presetColors={brushColors}
+								text="Brush"
+							/>
+							<ColorPicker
+								color={fillColor}
+								colorHandler={setFillColor}
+								disabled={disabled}
+								presetColors={fillColors}
+								text="Fill"
+							/>
+							<ColorPicker
+								color={canvasColor}
+								colorHandler={setCanvasColor}
+								disabled={disabled}
+								presetColors={canvasColors}
+								text="Canvas"
+							/>
+						</Cell>
 					</Cell>
 					<Cell>
 						<Row>
 							<UiDrawing
 								{...rest}
 								backgroundImage={backgroundImage}
-								brushColor={brushColorValue}
-								brushSize={brushSizeValue}
-								canvasColor={canvasColorValue}
+								brushColor={brushColor}
+								brushSize={brushSize}
+								canvasColor={canvasColor}
 								canvasHeight={canvasHeight}
 								canvasWidth={canvasWidth}
 								disabled={disabled}
-								drawingTool={drawingToolValue}
-								fillColor={fillColorValue}
+								drawingTool={drawingTool}
+								fillColor={fillColor}
 								ref={drawingRef}
 							/>
 						</Row>
 					</Cell>
 					<Cell shrink size="10%">
 						<Column align="center space-between" className={css.canvasOptions}>
-							{showDrawingUtils ? (
-								<ComponentOverride
-									backgroundImage={backgroundImage}
-									component={drawingUtilsComponent}
-									disabled={disabled}
-									drawingRef={drawingRef}
-									fileInputHandler={fileInputHandler}
-									setBackgroundImage={setBackgroundImage}
-								/>
-							) :
-								null
-							}
+							<Button css={css} disabled={disabled} icon="arrowhookleft" onClick={() => drawingRef.current.undo()} size="small" tooltipText="Undo" />
+							<Button css={css} disabled={disabled} icon="arrowhookright" onClick={() => drawingRef.current.redo()} size="small" tooltipText="Redo" />
+							<Button css={css} disabled={disabled} icon="refresh" onClick={() => drawingRef.current.clearCanvas()} size="small" tooltipText="Clear all" />
+							<Button css={css} disabled={disabled} icon="plus" onClick={() => document.getElementById('fileInput').click()} size="small" tooltipText="Import image" />
+							<input
+								accept="image/*"
+								className={css.inputFile}
+								id="fileInput"
+								onChange={(ev) => fileInputHandler({backgroundImage, ev, setBackgroundImage})}
+								onClick={(e) => {
+									e.target.value = null;
+								}}
+								type="file"
+							/>
+							<Button css={css} disabled={disabled} icon="trash" onClick={() => setBackgroundImage(null)} size="small" tooltipText="Clear image" />
+							<Button css={css} disabled={disabled} icon="download" onClick={() => drawingRef.current.saveCanvas()} size="small" tooltipText="Save canvas" />
 						</Column>
 					</Cell>
 				</Layout>
@@ -350,14 +265,7 @@ const DrawingDecorator = compose(
  *
  * Usage:
  * ```
- * <Drawing
- * 	brushColor="#545BCC"
- * 	brushSize={5}
- * 	canvasColor="#FFFFFF"
- * 	canvasHeight={800}
- * 	canvasWidth={1200}
- * 	drawingTool="brush"
- * />
+ * <Drawing />
  * ```
  *
  * @class Drawing
