--- Drawing/DrawingControls.module.less
@@ -1,25 +1,0 @@
-.drawingControls {
-	.bodyText {
-		font-size: 36px;
-		margin: 0;
-	}
-
-	.drawingTools {
-		margin: 0;
-
-		& > * {
-			font-size: 36px;
-		}
-	}
-
-	.slider {
-		margin: 0;
-		padding: 24px 0;
-	}
-
-	.toolbar {
-		.colors {
-			padding: 24px 0;
-		}
-	}
-}
