--- Drawing/DrawingUtils.module.less
@@ -1,10 +1,0 @@
-.drawingUtils {
-	.button {
-		min-width: 144px;
-		font-size: 36px;
-	}
-
-	.inputFile {
-		display: none;
-	}
-}
