+++ samples/sampler/stories/default/Drawing.js
@@ -1,7 +1,7 @@
 import BodyText from '@enact/sandstone/BodyText';
 import Drawing, {DrawingBase} from '@enact/sandstone/Drawing';
 import {mergeComponentMetadata} from '@enact/storybook-utils';
-import {boolean, number, range, select} from '@enact/storybook-utils/addons/controls';
+import {boolean, number} from '@enact/storybook-utils/addons/controls';
 import UIDrawing, {DrawingBase as UIDrawingBase} from '@enact/ui/Drawing';
 
 Drawing.displayName = 'Drawing';
@@ -13,6 +13,7 @@ const Config = mergeComponentMetadata(
 	Drawing
 );
 
+
 export default {
 	component: 'Drawing',
 	title: 'Sandstone/Drawing'
@@ -20,6 +21,8 @@ export default {
 
 export const _Drawing = (args) => {
 	const disabled = args['disabled'];
+	const canvasHeight = args['canvasHeight'];
+	const canvasWidth = args['canvasWidth'];
 
 	return (
 		<section>
@@ -27,31 +30,17 @@ export const _Drawing = (args) => {
 				<sup>*</sup>Drawing is not allowed while <code>disabled</code> is true.
 			</BodyText>
 			<Drawing
-				brushColor={args['brushColor']}
-				brushSize={args['brushSize']}
-				canvasColor={args['canvasColor']}
-				canvasHeight={args['canvasHeight']}
-				canvasWidth={args['canvasWidth']}
+				canvasHeight={canvasHeight}
+				canvasWidth={canvasWidth}
 				disabled={disabled}
-				drawingTool={args['drawingTool']}
-				fillColor={args['fillColor']}
-				showDrawingControls={args['showDrawingControls']}
-				showDrawingUtils={args['showDrawingUtils']}
 			/>
 		</section>
 	);
 };
 
-select('brushColor', _Drawing, ['', '#000000', '#FFFFFF', '#FF0000', '#00FF00'], Config, '');
-range('brushSize', _Drawing, Config, {min: 1, max: 30}, 5);
-select('canvasColor', _Drawing, ['', '#000000', '#FFFFFF', '#FF0000', '#00FF00'], Config, '');
 number('canvasHeight', _Drawing, Config, 800);
 number('canvasWidth', _Drawing, Config, 1200);
 boolean('disabled', _Drawing, Config);
-select('drawingTool', _Drawing, ['brush', 'fill', 'triangle', 'rectangle', 'circle', 'erase'], Config, '');
-select('fillColor', _Drawing, ['', '#000000', '#FFFFFF', '#FF0000', '#00FF00'], Config, '');
-boolean('showDrawingControls', _Drawing, Config);
-boolean('showDrawingUtils', _Drawing, Config);
 
 _Drawing.storyName = 'Drawing';
 _Drawing.parameters = {
